Fix `make clusters using=ovn`
a1ab011f2271478a694aeba864099f09f37ff93d broke
being able to specify a cable driver because
the setting were populated before the `using` var
was parsed from the command line

This commit ensures we choose the correct settings
yaml based on what is specified in the `using` flag.
By parsing it both in `Makefile` and `Makefile.inc` to
ensure it is always up to date.

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>